+++ syslog_process.php
@@ -23,7 +23,7 @@
 */
 
 include(dirname(__FILE__) . '/../../include/cli_check.php');
-include(dirname(__FILE__) . '/config.php');
+include(SYSLOG_CONFIG);
 include_once(dirname(__FILE__) . '/functions.php');
 
 /* Let it run for an hour if it has to, to clear up any big