+++ syslog_removal.php
@@ -109,7 +109,7 @@ function form_save() {
 function form_actions() {
 	global $config, $syslog_actions, $fields_syslog_action_edit;
 
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 
 	get_filter_request_var('drp_action', FILTER_VALIDATE_REGEXP,
 		 array('options' => array('regexp' => '/^([a-zA-Z0-9_]+)$/')));
@@ -247,7 +247,7 @@ function form_actions() {
 }
 
 function removal_export() {
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 
 	/* if we are to save this form, instead of display it */
 	if (isset_request_var('selected_items')) {
@@ -280,7 +280,7 @@ function removal_export() {
 function api_syslog_removal_save($id, $name, $type, $message, $rmethod, $notes, $enabled) {
 	global $config;
 
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 
 	/* get the username */
 	$username = db_fetch_cell('SELECT username FROM user_auth WHERE id=' . $_SESSION['sess_user_id']);
@@ -316,17 +316,17 @@ function api_syslog_removal_save($id, $name, $type, $message, $rmethod, $notes, 
 }
 
 function api_syslog_removal_remove($id) {
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 	syslog_db_execute("DELETE FROM `" . $syslogdb_default . "`.`syslog_remove` WHERE id='" . $id . "'");
 }
 
 function api_syslog_removal_disable($id) {
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 	syslog_db_execute("UPDATE `" . $syslogdb_default . "`.`syslog_remove` SET enabled='' WHERE id='" . $id . "'");
 }
 
 function api_syslog_removal_enable($id) {
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 	syslog_db_execute("UPDATE `" . $syslogdb_default . "`.`syslog_remove` SET enabled='on' WHERE id='" . $id . "'");
 }
 
@@ -349,7 +349,7 @@ function api_syslog_removal_reprocess($id) {
    --------------------- */
 
 function syslog_get_removal_records(&$sql_where, $rows) {
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 
 	if (get_request_var('filter') != '') {
 		$sql_where .= (strlen($sql_where) ? ' AND ':'WHERE ') .
@@ -383,7 +383,7 @@ function syslog_get_removal_records(&$sql_where, $rows) {
 function syslog_action_edit() {
 	global $message_types;
 
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 
 	/* ================= input validation ================= */
 	get_filter_request_var('id');
@@ -612,7 +612,7 @@ function syslog_filter() {
 function syslog_removal() {
 	global $syslog_actions, $message_types, $config;
 
-	include(dirname(__FILE__) . '/config.php');
+	include(SYSLOG_CONFIG);
 
     /* ================= input validation and session storage ================= */
     $filters = array(